Fix for bug in Update API of the ephemeral key recipe.
There was a problem in the way Update API was trying to update
the value of ephemeral key with new value. The problem was that
the Create API was calling a helper routine called RenewTTL which
was caching the original value of the key. So, when the user
calls Update API the key was getting updated with newer value
but was quickly getting overwritten by the original value cached
by the RenewTTL method. This is overcome by implementing the TTL
renewal logic in Create API and sending the newer value via a
channel which would be read by the go-routine created by Create
API for renewal.